Attempting to re-split out Email/SMS/Slack services using CommunicationsService as a base
