+++ src/main/java/org/eaa690/aerie/service/CommunicationService.java
@@ -16,58 +16,43 @@
 
 package org.eaa690.aerie.service;
 
-import java.io.IOException;
-import java.time.Instant;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
-import com.sendgrid.Method;
-import com.sendgrid.Request;
-import com.sendgrid.Response;
 import com.sendgrid.SendGrid;
-import com.sendgrid.helpers.mail.Mail;
-import com.sendgrid.helpers.mail.objects.Content;
-import com.sendgrid.helpers.mail.objects.Email;
-import com.sendgrid.helpers.mail.objects.Personalization;
 import com.ullink.slack.simpleslackapi.SlackSession;
-import com.ullink.slack.simpleslackapi.events.SlackMessagePosted;
-import com.ullink.slack.simpleslackapi.listeners.SlackMessagePostedListener;
+import lombok.Getter;
 import org.eaa690.aerie.communication.AcceptsEmailPredicate;
 import org.eaa690.aerie.communication.AcceptsSMSPredicate;
 import org.eaa690.aerie.communication.AcceptsSlackPredicate;
-import org.eaa690.aerie.constant.PropertyKeyConstants;
 import org.eaa690.aerie.exception.ResourceNotFoundException;
 import org.eaa690.aerie.model.Member;
 import org.eaa690.aerie.model.MemberRepository;
-import org.eaa690.aerie.model.MessageType;
 import org.eaa690.aerie.model.QueuedMessage;
 import org.eaa690.aerie.model.QueuedMessageRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.scheduling.annotation.Scheduled;
-import org.springframework.stereotype.Service;
 
 /**
  * CommunicationService.
  */
-@Service
-public class CommunicationService implements SlackMessagePostedListener {
+@Getter
+public abstract class CommunicationService {
+
+    /**
+     * Logger.
+     */
+    private static final Logger LOGGER = LoggerFactory.getLogger(CommunicationService.class);
 
     /**
      * SendGrid API.
      */
     @Autowired
     private SendGrid sendGrid;
-
-    /**
-     * Logger.
-     */
-    private static final Logger LOGGER = LoggerFactory.getLogger(CommunicationService.class);
 
     /**
      * PropertyService.
@@ -178,11 +163,7 @@ public class CommunicationService implements SlackMessagePostedListener {
      * @param queuedMessage QueuedMessage
      */
     public void queueMsg(final QueuedMessage queuedMessage) {
-        if (queuedMessage.getMessageType() == MessageType.Slack) {
-            sendSlackMessage(queuedMessage);
-        } else {
-            queuedMessageRepository.save(queuedMessage);
-        }
+        queuedMessageRepository.save(queuedMessage);
     }
 
     /**
@@ -197,48 +178,7 @@ public class CommunicationService implements SlackMessagePostedListener {
     /**
      * Looks for any messages in the send queue, and sends up to X (see configuration) messages per day.
      */
-    @Scheduled(cron = "0 0 10 * * *")
-    public void processQueue() {
-        final Optional<List<QueuedMessage>> allQueuedMessages = queuedMessageRepository.findAll();
-        if (allQueuedMessages.isPresent()) {
-            try {
-                allQueuedMessages
-                        .get()
-                        .stream()
-                        .limit(Long.parseLong(
-                                propertyService.get(PropertyKeyConstants.SEND_GRID_LIMIT).getValue()))
-                        .forEach(qe -> {
-                            sendSlackMessage(qe);
-                            sendSMSMessage(qe);
-                            sendEmailMessage(qe);
-                            queuedMessageRepository.delete(qe);
-                        });
-            } catch (ResourceNotFoundException e) {
-                LOGGER.error("Error", e);
-            }
-        }
-    }
-
-    /**
-     * Processes messages received by membership slack bot.
-     *
-     * @param event SlackMessagePosted
-     * @param session SlackSession
-     */
-    @Override
-    public void onEvent(final SlackMessagePosted event, final SlackSession session) {
-        // Ignore bot user messages
-        if (session.sessionPersona().getId().equals(event.getSender().getId())) {
-            return;
-        }
-        final String message = event.getMessageContent();
-        final String user = event.getUser().getUserName();
-        final String msg = String.format(
-                "Slack message received: user [%s]; message [%s]",
-                user,
-                message);
-        LOGGER.info(msg);
-    }
+    public abstract void processQueue();
 
     /**
      * Gets all Slack users.
@@ -260,7 +200,7 @@ public class CommunicationService implements SlackMessagePostedListener {
      * @param msgKey message key
      * @return message
      */
-    private String getSMSOrSlackMessage(final Member member, final String msgKey) {
+    protected String getSMSOrSlackMessage(final Member member, final String msgKey) {
         try {
             final String expiration;
             if (member.getExpiration() != null) {
@@ -283,204 +223,24 @@ public class CommunicationService implements SlackMessagePostedListener {
     }
 
     /**
-     * Sends a message via the Slack bot.
+     * Sends a message.
      *
      * @param queuedMessage QueuedMessage
      */
-    private void sendSlackMessage(final QueuedMessage queuedMessage) {
-        if (queuedMessage.getMessageType() == MessageType.Slack) {
-            Optional<Member> memberOpt = memberRepository.findById(queuedMessage.getMemberId());
-            if (memberOpt.isPresent() && acceptsSlackPredicate.test(memberOpt.get())) {
-                LOGGER.info(String.format("Sending Slack message [%s] to [%s]", queuedMessage.getBody(),
-                        queuedMessage.getRecipientAddress()));
-                slackSession.sendMessageToUser(slackSession.findUserByUserName(queuedMessage.getRecipientAddress()),
-                        queuedMessage.getBody(), null);
-            }
-        }
-    }
-
-    /**
-     * Sends an SMS message via SendGrid.
-     *
-     * @param queuedMessage QueuedMessage
-     */
-    private void sendSMSMessage(final QueuedMessage queuedMessage) {
-        if (queuedMessage.getMessageType() == MessageType.SMS) {
-            final Optional<Member> memberOpt = memberRepository.findById(queuedMessage.getMemberId());
-            if (memberOpt.isPresent()) {
-                final Member member = memberOpt.get();
-                if (acceptsSMSPredicate.test(member)) {
-                    queuedMessage.setRecipientAddress(String.format("%s@%s",
-                            queuedMessage.getRecipientAddress(),
-                            member.getCellPhoneProvider().getCellPhoneProviderEmailDomain()));
-                    sendEmailMessage(queuedMessage);
-                }
-            }
-        }
-    }
-
-    /**
-     * Sends an Email message via SendGrid.
-     *
-     * @param queuedMessage QueuedMessage
-     */
-    private void sendEmailMessage(final QueuedMessage queuedMessage) {
-        if (queuedMessage.getMessageType() == MessageType.Email) {
-            try {
-                final Optional<Member> memberOpt = memberRepository.findById(queuedMessage.getMemberId());
-                if (memberOpt.isPresent()) {
-                    final Member member = memberOpt.get();
-                    if (acceptsEmailPredicate.test(member)) {
-                        Email from = new Email(propertyService.get(PropertyKeyConstants.SEND_GRID_FROM_ADDRESS_KEY)
-                                .getValue());
-                        String subject = propertyService.get(queuedMessage.getSubjectKey()).getValue();
-                        Email to = new Email(queuedMessage.getRecipientAddress());
-                        if (Boolean.parseBoolean(
-                                propertyService.get(PropertyKeyConstants.EMAIL_TEST_MODE_ENABLED_KEY).getValue())) {
-                            to = new Email(propertyService.get(PropertyKeyConstants.EMAIL_TEST_MODE_RECIPIENT_KEY)
-                                    .getValue());
-                        }
-
-                        final Mail mail;
-                        if (queuedMessage.getTemplateIdKey() != null) {
-                            mail = new Mail();
-                            mail.setSubject(subject);
-                            mail.setTemplateId(propertyService.get(queuedMessage.getTemplateIdKey()).getValue());
-                            mail.addPersonalization(personalize(member, to));
-                            mail.setFrom(from);
-                        } else {
-                            mail = new Mail(from, subject, to, new Content("text/plain",
-                                    queuedMessage.getBody()));
-                        }
-
-                        final Request request = new Request();
-                        request.setMethod(Method.POST);
-                        request.setEndpoint("mail/send");
-                        request.setBody(mail.build());
-                        Response response = sendGrid.api(request);
-                        LOGGER.info(String.format("Response... statusCode [%s]; body [%s]; headers [%s]",
-                                response.getStatusCode(), response.getBody(), response.getHeaders()));
-                    }
-                }
-            } catch (IOException | ResourceNotFoundException ex) {
-                LOGGER.error(ex.getMessage());
-            }
-        }
-    }
-
-    /**
-     * Personalizes an email to the member.
-     *
-     * @param member Member
-     * @param to address
-     * @return Personalization
-     * @throws ResourceNotFoundException when property is not found
-     */
-    private Personalization personalize(final Member member, final Email to) throws ResourceNotFoundException {
-        final Personalization personalization = new Personalization();
-        personalization.addTo(to);
-        personalization.addBcc(new Email(propertyService.get(PropertyKeyConstants.EMAIL_BCC_KEY).getValue()));
-        personalization.addDynamicTemplateData("firstName", member.getFirstName());
-        personalization.addDynamicTemplateData("lastName", member.getLastName());
-        personalization.addDynamicTemplateData("url", jotFormService.buildRenewMembershipUrl(member));
-        if (member.getExpiration() == null) {
-            personalization.addDynamicTemplateData("expirationDate", ZonedDateTime.ofInstant(Instant.now(),
-                    ZoneId.systemDefault()).format(
-                    DateTimeFormatter.ofPattern("MMM d, yyyy")));
-        } else {
-            personalization.addDynamicTemplateData("expirationDate", ZonedDateTime.ofInstant(
-                    member.getExpiration().toInstant(),
-                    ZoneId.systemDefault()).format(
-                    DateTimeFormatter.ofPattern("MMM d, yyyy")));
-        }
-        return personalization;
-    }
+    public abstract void sendMessage(QueuedMessage queuedMessage);
 
     /**
      * Sends new member message.
      *
      * @param member Member
      */
-    public void sendNewMembershipMsg(final Member member) {
-        if (member != null && member.getSlack() != null && member.isSlackEnabled()) {
-            try {
-                final QueuedMessage queuedSlackMessage = new QueuedMessage();
-                queuedSlackMessage.setMemberId(member.getId());
-                queuedSlackMessage.setBody(getSMSOrSlackMessage(member, PropertyKeyConstants.SLACK_NEW_MEMBER_MSG_KEY));
-                queuedSlackMessage.setMessageType(MessageType.Slack);
-                queuedSlackMessage.setRecipientAddress(getSlackName(member));
-                queueMsg(queuedSlackMessage);
-
-                final QueuedMessage queuedSMSMessage = new QueuedMessage();
-                queuedSMSMessage.setMemberId(member.getId());
-                queuedSMSMessage.setBody(getSMSOrSlackMessage(member, PropertyKeyConstants.SMS_NEW_MEMBER_MSG_KEY));
-                queuedSMSMessage.setMessageType(MessageType.SMS);
-                queuedSMSMessage.setRecipientAddress(member.getCellPhone());
-                queueMsg(queuedSMSMessage);
-
-                final QueuedMessage queuedEmailMessage = new QueuedMessage();
-                queuedEmailMessage.setMemberId(member.getId());
-                queuedEmailMessage.setSubjectKey(PropertyKeyConstants.SEND_GRID_NEW_MEMBERSHIP_EMAIL_SUBJECT_KEY);
-                queuedEmailMessage.setTemplateIdKey(PropertyKeyConstants.SEND_GRID_NEW_MEMBERSHIP_EMAIL_TEMPLATE_ID);
-                queuedEmailMessage.setRecipientAddress(member.getEmail());
-                queuedEmailMessage.setMessageType(MessageType.Email);
-                queueMsg(queuedEmailMessage);
-            } catch (ResourceNotFoundException ex) {
-                LOGGER.error(ex.getMessage());
-            }
-        }
-    }
+    public abstract void sendNewMembershipMsg(Member member);
 
     /**
      * Sends membership renewal message.
      *
      * @param member Member
      */
-    public void sendRenewMembershipMsg(final Member member) {
-        if (member != null && member.getSlack() != null && member.isSlackEnabled()) {
-            try {
-                final QueuedMessage queuedSlackMessage = new QueuedMessage();
-                queuedSlackMessage.setMemberId(member.getId());
-                queuedSlackMessage.setBody(getSMSOrSlackMessage(member,
-                        PropertyKeyConstants.SLACK_RENEW_MEMBER_MSG_KEY));
-                queuedSlackMessage.setMessageType(MessageType.Slack);
-                queuedSlackMessage.setRecipientAddress(getSlackName(member));
-                queueMsg(queuedSlackMessage);
-
-                final QueuedMessage queuedSMSMessage = new QueuedMessage();
-                queuedSMSMessage.setMemberId(member.getId());
-                queuedSMSMessage.setBody(getSMSOrSlackMessage(member, PropertyKeyConstants.SMS_RENEW_MEMBER_MSG_KEY));
-                queuedSMSMessage.setMessageType(MessageType.SMS);
-                queuedSMSMessage.setRecipientAddress(member.getCellPhone());
-                queueMsg(queuedSMSMessage);
-
-                final QueuedMessage queuedEmailMessage = new QueuedMessage();
-                queuedEmailMessage.setMemberId(member.getId());
-                queuedEmailMessage.setSubjectKey(PropertyKeyConstants.SEND_GRID_NEW_MEMBERSHIP_EMAIL_SUBJECT_KEY);
-                queuedEmailMessage.setTemplateIdKey(PropertyKeyConstants.SEND_GRID_NEW_MEMBERSHIP_EMAIL_TEMPLATE_ID);
-                queuedEmailMessage.setRecipientAddress(member.getEmail());
-                queuedEmailMessage.setMessageType(MessageType.Email);
-                queueMsg(queuedEmailMessage);
-            } catch (ResourceNotFoundException ex) {
-                LOGGER.error(ex.getMessage());
-            }
-        }
-    }
-
-    /**
-     * Gets Slack name, or overridden value, if in test mode.
-     *
-     * @param member Member
-     * @return Slack name
-     * @throws ResourceNotFoundException if properties are not found
-     */
-    private String getSlackName(final Member member) throws ResourceNotFoundException {
-        String to = member.getSlack();
-        if (Boolean.parseBoolean(
-                propertyService.get(PropertyKeyConstants.SLACK_TEST_MODE_ENABLED_KEY).getValue())) {
-            to = propertyService.get(PropertyKeyConstants.SLACK_TEST_MODE_RECIPIENT_KEY).getValue();
-        }
-        return to;
-    }
+    public abstract void sendRenewMembershipMsg(Member member);
 
 }
