+++ src/main/resources/application.yaml
@@ -17,10 +17,10 @@
 spring:
   datasource:
     username: aerie
-    password: ${AERIE_DB_PASS}
-    url: jdbc:mysql://localhost:3306/aerie?serverTimezone=America/New_York
+    password: aerie
+    url: jdbc:mysql://aerie-mysql:33061/aerie?serverTimezone=America/New_York
   jpa:
     database-platform: org.hibernate.dialect.MySQL57Dialect
-    show-sql: true
+    show-sql: false
     hibernate:
       ddl-auto: update
